oktaSdk destructured to its own middleware
